+++ src/aioway/__init__.py
@@ -10,10 +10,10 @@ from . import (
     _utils,
     api,
     app,
+    deductions,
     dsets,
     emits,
     errors,
-    instrs,
     modes,
     rewrites,
     schemas,
