+++ src/aioway/emits/convs.py
@@ -8,10 +8,10 @@ from torch import nn
 from torchrl.data import tensor_specs as tspecs
 
 from aioway._utils import is_list_of
-from aioway.instrs import Activation
 from aioway.schemas import Shape
 from aioway.tspecs import TSpec, sample_from_tspec, unbounded_box_tspec
 
+from ._utils import Activation
 from .emitters import Emitter, emitter_dcls
 from .linear import linear_regression
