--- src/aioway/instrs/__init__.py
@@ -1,9 +1,0 @@
-# Copyright (c) AIoWay Authors - All Rights Reserved
-
-"`Instr` is the IR for `aioway`."
-
-from .containers import *
-from .instrs import *
-from .layers import *
-from .nn import *
-from .rl import *
