--- src/aioway/instrs/layers/__init__.py
@@ -1,8 +1,0 @@
-# Copyright (c) AIoWay Authors - All Rights Reserved
-
-from .actives import *
-from .dense import *
-from .dropouts import *
-from .embs import *
-from .norms import *
-from .sliding import *
