+++ src/aioway/losses/__init__.py
@@ -1,6 +1,5 @@
 # Copyright (c) AIoWay Authors - All Rights Reserved
 
-from .bases import *
 from .contrastive import *
 from .emits import *
 from .recon import *
